pyproject.toml
[project]
name = "todo-reasoning-agent"
version = "0.1.0"
description = "Worker-Evaluator Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.0.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
]
